Create all the components
