--- src/components/Books.js
@@ -1,17 +1,0 @@
-import React from 'react';
-
-const Form = () => (
-  <return>
-    <form className="form-container">
-      <input
-        type="text"
-        className="input-text"
-        placeholder="Add book..."
-        name="title"
-      />
-      <button type="submit" className="input-submit">Submit</button>
-    </form>
-  </return>
-);
-
-export default Form;
