+++ src/components/List.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Book from './Book';
+
+const List = () => (
+  <div>
+    <Book title="The Hunger Games" author="Suzanne Collins" />
+    <Book title="Dune" author="Frank Herbert" />
+    <Book title="Capital in the Twenty-First Century" author="Suzanne Collins" />
+  </div>
+);
+
+export default List;
