+++ src/index.css
@@ -11,3 +11,24 @@ html {
 body {
   background-color: rgb(244, 244, 239);
 }
+
+.navBar {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-around;
+  align-items: center;
+  padding: 1rem;
+  height: 7rem;
+  width: 100%;
+}
+
+.menuNav {
+  display: flex;
+  list-style: none;
+  margin-right: 3rem;
+}
+
+.navList {
+  padding: 0.6rem 1.5rem;
+  margin: 0 1rem;
+}
