+++ src/pages/Home.js
@@ -1,8 +1,10 @@
 import React from 'react';
-import Form from '../components/Books';
+import Form from '../components/Form';
+import List from '../components/List';
 
 const Books = () => (
   <div>
+    <List />
     <Form />
   </div>
 );
